Cambios en TraerUsuario, añado funcionalidades - Ahora el metodo TraerUsuario devuelve el objeto UsuarioWS. Mejora en logica de busqueda con LINQ - Mejoras y cambios en logica de intentos de Login - Metodo LoginWS ahora devuelve el StatusCode como Int(string) en lugar del responseText - Se añade logica para sugerencias de usuarios logueados anteriormente - FormLogin: Cambios en nombres de metodos/variables, se añade logica para redireccion a pantalla segun Host (perfil) de usuario
To-Do:
-- Mejorar diseños de pantallas (a ser mostrados en ejemplo)
-- Empezar y organizar para comenzar con logica de los demas modulos

+++ TemplateTPIntegrador/Persistencia/LoginWS.cs
@@ -9,6 +9,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using System.Diagnostics;
+using System.Net;
 
 namespace Persistencia
 {
@@ -48,7 +49,7 @@ namespace Persistencia
                 throw new Exception($"Error: {responseBody}");
             }
 
-            return (idUsuario, responseBody, response.StatusCode.ToString());
+            return (idUsuario, responseBody, ((int)response.StatusCode).ToString());
         }
 
     }
